// database/types.ts
export interface LinkType {
  link: string;
  label: string;
}

// export interface JobOfferType {
//   user: string;
//   title: string;
//   description: string;
//   requirements: string[];
//   requireCount: number;
//   onSite: boolean;
//   location: string | null;
//   salary: number;
//   allowance: string[];
//   deadline: Date;
//   contact: LinkType[];
//   created: Date;
// }

// export interface ReplyType {
//   title: string;
//   description: string;
//   appointment: Date;
//   link: LinkType[];
//   created: Date;
// }

// export interface FreelanceOfferType {
//   user: string;
//   title: string;
//   detail: string;
//   requirements: string[];
//   responsibilities: string[];
//   salaryFrom: number; //From 0 $
//   salaryTo: number; // To 0 $
//   field: string[]; //web /software/ programming/ video-editing/ copy-writing
//   timeRange: string; // 1 day/2day /3 days/ 1 week/ 1 month/
//   contact: LinkType[];
//   formClose: string; // 1 day/2day /3 days/ 1 week/ 1 month/
// }

// export interface FreelanceApplicationType {
//   user: string;
//   title: string;
//   description: string;
//   cv: string; //file type
//   negoSalary: number; //
//   links: LinkType[]; //
//   approved: boolean;
//   created: Date;
// }

// export interface JobApplicationType {
//   user: string;
//   title: string;
//   description: string;
//   cv: string; //file type
//   links: LinkType[]; //
//   approved: boolean;
//   reply: ReplyType;
//   created: Date;
// }

// export interface UserType {
//   email: string;
//   password: string;
//   userProfile: string;
// }